+++ routes/postsRouter.js
@@ -1,22 +1,25 @@
 import { Router } from 'express';
-import validateJOI from '../middlewares/validateJOI.js';
+import validateZod from '../middlewares/validateZod.js';
 import {
-  createPost,
-  deletePost,
-  getAllPosts,
-  getSinglePost,
-  updatePost
+    createPost,
+    deletePost,
+    getAllPosts,
+    getSinglePost,
+    updatePost,
 } from '../controllers/posts.js';
-import { postSchema } from '../joi/schemas.js';
+import { postSchema } from '../zod/schemas.js';
 
 const postsRouter = Router();
 
-postsRouter.route('/').get(getAllPosts).post(validateJOI(postSchema), createPost);
+postsRouter
+    .route('/')
+    .get(getAllPosts)
+    .post(validateZod(postSchema), createPost);
 
 postsRouter
-  .route('/:id')
-  .get(getSinglePost)
-  .put(validateJOI(postSchema), updatePost)
-  .delete(deletePost);
+    .route('/:id')
+    .get(getSinglePost)
+    .put(validateZod(postSchema), updatePost)
+    .delete(deletePost);
 
 export default postsRouter;
